+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Fix typo in `assignment` callbaks
+* Alias `assignment` to `assigment` for backwards compatibility
 ## Version 0.2.0
 
 * Don't expose model name on `model`. _(security fix)_